Implementing a full HTTP/2-over-TCP proxy.
This adds two stub implementations

- ``display()`` for generic "printing", but this centralizes all such
  calls so they can be swapped at a later date if need be; e.g. for
  calls to write to a file or use ``logging``
- ``describe()`` to parse and describe an HTTP/2 frame; this just
  displays it for now

There are a few places where this proxy "cuts corners" and isn't as
robust as it could be. These are documented where possible; e.g. in
``recv()`` we say that we just throw a cowardly exception if the TCP
chunk is the same size that we read.